requirements.txt
streamlit==1.39.0
litellm==1.50.0
replicate==1.0.4
google-genai>=1.3.0
moviepy==1.0.3
python-dotenv==1.0.1
tqdm==4.66.5

